refactor(node): move mutationCLI into mutation-cli.ts (#246)
The orchestrator was in `src/mutation/index.ts`; put it in a file named for what it
is — `src/mutation/mutation-cli.ts` exposes `mutationCLI`, and the executable
`main.ts` imports it. The spawned entry (`dist/mutation/main.js`) is unchanged, so
nothing outside the folder moves.
