+++ packages/node/src/mutation/main.ts
@@ -1,9 +1,9 @@
-import { mutationCLI } from './index.js';
+import { mutationCLI } from './mutation-cli.js';
 
 // The executable the rust binary spawns for the TS mutation arm (#246): `dist/mutation/main.js`
 // runs `mutationCLI` over the process arguments and maps a failed run onto a non-zero exit code.
-// Kept separate from `index.ts` so the orchestration stays a pure, importable function with no
-// process side effects (and each stays fully covered).
+// Kept separate from `mutation-cli.ts` so the orchestration stays a pure, importable function with
+// no process side effects (and each stays fully covered).
 mutationCLI(process.argv.slice(2)).catch((err: Error) => {
   process.stderr.write(`${err.message}\n`);
   process.exitCode = 1;
